pyproject.toml
[project]
name = "telegram-transcriber-bot"
version = "0.1.0"
description = "Telegram bot that transcribes links and uploaded media, then generates research reports via cglm."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.22.0",
  "faster-whisper>=1.2.0",
  "python-docx>=1.2.0",
  "python-dotenv>=1.1.1",
  "youtube-transcript-api>=1.2.3",
  "yt-dlp>=2025.1.15",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "pytest-asyncio>=1.2.0",
  "pytest-cov>=7.0.0",
]

[project.scripts]
telegram-transcriber-bot = "telegram_transcriber_bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/telegram_transcriber_bot"]

[tool.pytest.ini_options]
pythonpath = ["workers/common/src", "src"]
testpaths = ["tests"]
addopts = [
  "-ra",
  "--strict-config",
  "--strict-markers",
  "--cov=transcriber_workers_common",
  "--cov-branch",
  "--cov-report=term-missing",
  "--cov-fail-under=95",
]
